Ensure Agent.Setup on .NET FullFramework behaves the same
Ensures that on .NET FullFramework we always use ConfigurationManager
