+++ src/integrations/Elastic.Apm.AspNetFullFramework/AgentComponentsUsingHttpContext.cs
@@ -1,0 +1,32 @@
+// Licensed to Elasticsearch B.V under
+// one or more agreements.
+// Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
+// See the LICENSE file in the project root for more information
+#nullable enable
+using Elastic.Apm.Config;
+using Elastic.Apm.Logging;
+
+namespace Elastic.Apm.AspNetFullFramework
+{
+	/// <summary>
+	/// Agent components for ASP.NET Full Framework
+	/// </summary>
+	internal class AgentComponentsUsingHttpContext : AgentComponents
+	{
+		public AgentComponentsUsingHttpContext(
+			IApmLogger logger,
+			IConfigurationReader configurationReader
+		)
+			: base(
+				logger,
+				configurationReader,
+				null,
+				null,
+				// ,
+				new HttpContextCurrentExecutionSegmentsContainer(),
+				null,
+				null)
+		{
+		}
+	}
+}
